Hotfix a read failure related to unmatched UTF16 surrogate pairs
Java Strings are complicated:
They allow unmatched surrogate pairs.

Rust strings do not allow this.

The old code was more correct, but failed on unmatched surrogates. Swapping the if cases around results in more use of the replacement character, but doesn't lead to a "crash".

So we did this, but further work is still needed, as currently:
- written class files can contain strings different from the read ones.